+++ bindings/angular2/src/ngx-onsenui.ts
@@ -27,6 +27,7 @@ import {OnsSwitch} from './directives/ons-switch';
 import {OnsRange} from './directives/ons-range';
 import {OnsSelect} from './directives/ons-select';
 import {OnsInput} from './directives/ons-input';
+import {OnsSearchInput} from './directives/ons-search-input';
 import {OnsPullHook} from './directives/ons-pull-hook';
 import {OnsLazyRepeat} from './directives/ons-lazy-repeat';
 import {OnsSplitterSide, OnsSplitterContent} from './directives/ons-splitter';
@@ -44,6 +45,7 @@ const directives = [
   OnsRange,
   OnsSelect,
   OnsInput,
+  OnsSearchInput,
   OnsPullHook,
   OnsLazyRepeat,
   OnsSplitterSide,